Refactor `anvil-types-<name>`
