+++ packages/anvil-plugin-babel/package.json
@@ -23,6 +23,8 @@
   "dependencies": {
     "@babel/core": "^7.1.2",
     "@babel/preset-env": "^7.1.0",
+    "@financial-times/anvil-types-adonai": "file:../anvil-types-adonai",
+    "@financial-times/anvil-types-build": "file:../anvil-types-build",
     "adonai": "^0.1.3",
     "babel-loader": "^8.0.4",
     "webpack-merge": "^4.1.4"
