+++ packages/anvil-plugin-css/package.json
@@ -21,7 +21,7 @@
     "@financial-times/anvil": "file:../anvil"
   },
   "dependencies": {
-    "@financial-times/anvil-plugin-helpers": "file:../anvil-plugin-helpers",
+    "@financial-times/anvil-types-build": "file:../anvil-types-build",
     "adonai": "^0.1.3",
     "css-loader": "^1.0.1",
     "style-loader": "^0.23.1"
