+++ packages/anvil-plugin-esnext/package.json
@@ -25,6 +25,8 @@
     "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
     "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@financial-times/anvil-types-adonai": "file:../anvil-types-adonai",
+    "@financial-times/anvil-types-build": "file:../anvil-types-build",
     "adonai": "^0.1.3"
   }
 }
