+++ packages/anvil-plugin-esnext/src/index.ts
@@ -1,10 +1,11 @@
 import { Plugin } from 'adonai'
 import babelPreset from './babel'
+import { RunningBabelContext } from '@financial-times/anvil-types-build'
 
 export default new Plugin(({ on }) => {
   on('@Build::amend::babelConfig', amendBabelConfig)
 })
 
-function amendBabelConfig({ c, babelConfig }) {
+function amendBabelConfig({ c, babelConfig }: RunningBabelContext) {
   babelConfig.presets.push(babelPreset(c))
 }
