+++ packages/anvil-plugin-ft-js/src/index.ts
@@ -1,5 +1,6 @@
 import { Plugin } from 'adonai'
 import babelPreset from './babel'
+import { RunningBabelContext } from '@financial-times/anvil-types-build'
 
 export default new Plugin(({ on }) => {
   on('@Build::amend::babelConfig', addPresetToBabelConfig)
@@ -13,7 +14,7 @@ function useCacheDirInScriptsRule({ scriptsRule }) {
   scriptsRule.use.options.cacheDirectory = true
 }
 
-function addPresetToBabelConfig({ c, babelConfig }) {
+function addPresetToBabelConfig({ c, babelConfig }: RunningBabelContext) {
   babelConfig.presets.push(babelPreset(c))
 }
