--- packages/anvil-plugin-helpers/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@financial-times/anvil-plugin-helpers",
-  "version": "0.0.0",
-  "description": "",
-  "main": "dist/cjs/index.js",
-  "types": "src/index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "tsc": "../../node_modules/.bin/tsc",
-    "clean": "npm run clean:dist && npm run clean:node_modules",
-    "clean:dist": "rm -rf dist",
-    "clean:node_modules": "rm -rf node_modules",
-    "build:cjs": "npm run tsc -- --outDir ./dist/cjs",
-    "build": "npm run clean:dist && npm run build:cjs",
-    "dev": "npm run clean:dist && npm run build:cjs -- --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "adonai": "^0.1.3",
-    "cli-progress": "^2.1.0",
-    "find-up": "^3.0.0",
-    "inquirer": "^6.2.0",
-    "ora": "^3.0.0"
-  }
-}
