--- packages/anvil-plugin-helpers/src/index.ts
@@ -1,3 +1,0 @@
-export * from './cli/CliContext'
-export * from './cli/CliMessenger'
-export * from './path/getPackageFolderPathRelativeTo'
