+++ packages/anvil-types-adonai/README.md
@@ -1,0 +1,9 @@
+# @financial-times/anvil-types-adonai
+
+Adonai specific type and interface definitions
+
+## Installation
+
+```
+npm install --save-dev @financial-times/anvil-types-generic
+```
