+++ packages/anvil-types-adonai/index.d.ts
@@ -1,0 +1,7 @@
+export interface Amender {
+  (hook: string, value: any): any
+}
+
+export interface OptionalAmender {
+  amend?: Amender
+}
