+++ packages/anvil-types-build/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@financial-times/anvil-types-build",
+  "version": "0.0.0",
+  "description": "",
+  "main": "index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "clean": "npm run clean:node_modules",
+    "clean:node_modules": "rm -rf node_modules"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@financial-times/anvil": "file:../anvil",
+    "@financial-times/anvil-types-generic": "file:../anvil-types-generic"
+  }
+}
