+++ packages/anvil-types-generic/.npmignore
@@ -1,0 +1,1 @@
+__test__/
