+++ packages/anvil-types-generic/.npmrc
@@ -1,0 +1,2 @@
+package-lock=false
+shrinkwrap=false
