+++ packages/anvil-types-generic/README.md
@@ -1,0 +1,9 @@
+# @financial-times/anvil-types-generic
+
+Generic type and interface definitions that are common across packages
+
+## Installation
+
+```
+npm install --save-dev @financial-times/anvil-types-generic
+```
