+++ packages/anvil-types-generic/index.d.ts
@@ -1,0 +1,3 @@
+export interface AnyObject {
+  [key: string]: any
+}
