+++ packages/anvil-utils/.npmignore
@@ -1,0 +1,1 @@
+__test__/
