+++ packages/anvil-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@financial-times/anvil-webpack",
+  "name": "@financial-times/anvil-utils",
   "version": "0.0.0",
   "description": "",
   "main": "dist/cjs/index.js",
@@ -18,10 +18,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/core": "^7.1.2",
-    "babel-loader": "^8.0.4",
-    "clean-webpack-plugin": "^0.1.19",
-    "@financial-times/anvil-plugin-helpers": "file:../anvil-plugin-helpers",
-    "webpack": "^4.23.1"
+    "find-up": "^3.0.0"
   }
 }
