--- packages/anvil-webpack/src/index.ts
@@ -1,2 +1,0 @@
-export * from './utils/pack'
-export * from './utils/getDefaultWebpackConfig'
