+++ packages/anvil/package.json
@@ -2,6 +2,8 @@
   "name": "@financial-times/anvil",
   "version": "0.0.0",
   "description": "The anvil CLI",
+  "main": "dist/cjs/index.js",
+  "types": "src/index.ts",
   "bin": {
     "anvil": "bin/cli.js"
   },
@@ -19,9 +21,16 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@financial-times/anvil-plugin-helpers": "file:../anvil-plugin-helpers",
-    "@financial-times/anvil-webpack": "file:../anvil-webpack",
+    "@financial-times/anvil-utils": "file:../anvil-utils",
     "adonai": "^0.1.3",
-    "commander": "^2.19.0"
+    "clean-webpack-plugin": "^0.1.19",
+    "cli-progress": "^2.1.0",
+    "commander": "^2.19.0",
+    "inquirer": "^6.2.0",
+    "ora": "^3.0.0",
+    "webpack": "^4.23.1"
+  },
+  "devDependencies": {
+    "@financial-times/anvil-types-generic": "file:../anvil-types-generic"
   }
 }
