+++ packages/anvil/src/index.ts
@@ -1,1 +1,2 @@
 export * from './program/createProgram'
+export * from './context/CliContext'
