+++ packages/anvil/src/operations/getDefaultWebpackConfig.ts
@@ -1,4 +1,4 @@
-import { CliContext } from '@financial-times/anvil-plugin-helpers'
+import { CliContext } from '../context/CliContext'
 import CleanWebpackPlugin from 'clean-webpack-plugin'
 
 export function getDefaultWebpackConfig(c: CliContext) {
