--- types/SharedTypes.d.ts
@@ -1,3 +1,0 @@
-interface AnyObject {
-  [key: string]: any
-}
